scholarqa/requirements.txt
# Core dependencies
python-dotenv==1.0.0

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# PDF Processing
pypdf==3.17.4
pdfplumber==0.10.3

# Text Processing & Embeddings
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.1.2

# Vector Database (Local)
chromadb==0.4.22

# LLM Local Models
llama-cpp-python==0.2.27

# Utilities
numpy==1.24.3
pandas==2.1.4
tqdm==4.66.1
psutil==5.9.6

# Text Splitting & Processing
tiktoken==0.5.2
nltk==3.8.1

# API & HTTP
requests==2.31.0
urllib3==2.1.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
